+++ CHANGELOG.md
@@ -8,8 +8,15 @@
 - Breaking changes to the `createConfig()` argument (object):
   - property `wrongMethodBehavior` (number) changed to `hintAllowedMethods` (boolean);
   - property `methodLikeRouteBehavior` (string literal) changed to `recognizeMethodDependentRoutes` (boolean);
+- Breaking change to the `EndpointsFactory::build()` argument (object):
+  - property `shortDescription` renamed to `summary`;
 - Breaking change to the `Documentation` constructor argument (object):
-  - property `hasSummaryFromDescription` renamed to `hasSummary`;
+  - property `hasSummaryFromDescription` (boolean) replaced with `summarizer` (function);
+  - If used with `false` value, replace it with `summarizer: ({ summary, trim }) => trim(summary)` for same behavior;
+- Featuring `summarizer` option to customize the summary of the Endpoint in the generated Documentation:
+  - The function receives `summary`, `description` and the default `trim()` function as arguments;
+  - The default summarizer uses `description` as a fallback for missing `summary`;
+  - The `trim()` function accepts a string and the limit (default: 50, best practice) that you can now customize;
 - Breaking change to the `Integration` constructor argument (object):
   - property `noContent` renamed to `noBodySchema`;
 - Consider using [the automated migration](https://www.npmjs.com/package/@express-zod-api/migration).
@@ -24,9 +31,16 @@
 ```
 
 ```diff
+  factory.build({
+-   shortDescription: "Retrieves the user.",
++   summary: "Retrieves the user.",
+  });
+```
+
+```diff
   new Documentation({
 -   hasSummaryFromDescription: false,
-+   hasSummary: false,
++   summarizer: ({ summary, trim }) => trim(summary),
   });
 ```